Implement comparison methods: '<', '>', 'in'
In `Interval`, `DaytimePeriod` and `Season`, the functions have been
implemented as far as it makes sense:

__lt__, __gt__, __contains__

The __contains__ method has a wide applicability. Interval,
DaytimePeriod and Season can be mutually compared whether they contain
eachother; with the exception of DaytimePeriod/Season (due to lack of
intuitive interpretation). Comparison to Iterables is also possible in
each of these cases.